rc5

+++ scripts/tcl_commands/synthesis.tcl
@@ -37,25 +37,40 @@ proc run_yosys {args} {
 	set ::env(SAVE_NETLIST) $::env(yosys_result_file_tag).v
     }
 
-    try_catch [get_yosys_bin] \
-	-c $::env(SYNTH_SCRIPT) \
-	-l $::env(yosys_log_file_tag).log \
-	|& tee $::env(TERMINAL_OUTPUT)
+	if { [file exists $::env(SAVE_NETLIST)] } {
+		puts_warn "A netlist at $::env(SAVE_NETLIST) already exists..."
+		puts_warn "Skipping synthesis"
+	} else {
+		try_catch [get_yosys_bin] \
+			-c $::env(SYNTH_SCRIPT) \
+			-l $::env(yosys_log_file_tag).log \
+			|& tee $::env(TERMINAL_OUTPUT)
+	}
 
     set_netlist $::env(SAVE_NETLIST)
     if { $::env(LEC_ENABLE) && [file exists $::env(PREV_NETLIST)] } {
 	logic_equiv_check -rhs $::env(PREV_NETLIST) -lhs $::env(CURRENT_NETLIST)
     }
 
-
+    # The following is a naive workaround to the defparam issue.. it should be handled with
+    # an issue to the OpenROAD verilog parser.
+    if { [info exists ::env(SYNTH_EXPLORE)] && $::env(SYNTH_EXPLORE) } {
+        puts_info "This is a Synthesis Exploration and so no need to remove the defparam lines."
+    } else {
+        try_catch sed -ie {/defparam/d} $::env(CURRENT_NETLIST)
+    }
     TIMER::timer_stop
     exec echo "[TIMER::get_runtime]" >> $::env(yosys_log_file_tag)_runtime.txt
 }
 
 proc run_sta {args} {
     puts_info "Running Static Timing Analysis..."
-    try_catch sta $::env(SCRIPTS_DIR)/sta.tcl \
-	|& tee $::env(TERMINAL_OUTPUT) $::env(opensta_log_file_tag).log
+    if {[info exists ::env(CLOCK_PORT)]} {
+        try_catch sta $::env(SCRIPTS_DIR)/sta.tcl \
+        |& tee $::env(TERMINAL_OUTPUT) $::env(opensta_log_file_tag).log
+    } else {
+        puts_warn "No CLOCK_PORT found. Skipping STA..."
+    }
 }
 
 proc run_synth_exploration {args} {
